修正

+++ db/migrate/20200428132411_create_sending_destinations.rb
@@ -10,7 +10,7 @@ class CreateSendingDestinations < ActiveRecord::Migration[5.2]
       t.string :city, null: false
       t.string :house_number, null: false
       t.string :building_name
-      t.integer :phone_number, unique: true
+      t.bigint :phone_number, unique: true
       t.references :user, null: false, foreign_key: true
       t.timestamps
     end
